+++ src/pipelines/features/PipelineTemplatesTable/PipelineTemplateTable.generated.tsx
